Revert "merge master into vue rewrite"
