+++ .gitignore
@@ -2,5 +2,3 @@
 /node_modules
 exports/barcode-data.json
 builds
-exports
-barcode_data.tab
