+++ barcode-search.html
@@ -1,0 +1,63 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+        <link rel="stylesheet" href="node_modules/xel/stylesheets/material.theme.css">
+        <script src="node_modules/xel/xel.min.js"></script>
+        <link href="https://fonts.googleapis.com/css?family=Roboto|Roboto+Mono" rel="stylesheet">
+</head>
+
+<body OnLoad="document.getElementById('search-box').focus();">
+        <x-drawer id="drawer">
+            <x-button id="openFileButton">
+                <x-label>Load Data</x-label>
+              </x-button>
+            </x-drawer>
+        
+        <x-button id="button">
+          <x-label>Menu</x-label>
+        </x-button>
+        
+        <style>
+          #drawer {
+            width: 30vw;
+            height: 70vh;
+            color: white;
+            background: #009788;
+            border-radius: 0 5px 5px 0;
+            opacity: 0.85;
+            font-size: 16px;
+            padding: 10px 20px;
+          }
+        </style>
+        
+        <script>
+          let drawer = document.querySelector("#drawer");
+          let button = document.querySelector("#button");
+          button.addEventListener("click", () => drawer.opened = true);
+        </script>
+    <br /> Sample Barcode for testing:
+    <div id="sample-barcode">0055122673</div>
+    <input type="button" class="sample-code" id="test-sample-code" value="Search using test barcode" onclick="test_func()">
+    <div id="data-file"></div>
+    <div id="search-inputs">
+            <x-box>
+        <x-input class="search-code" id="search-box">
+        <x-label>Enter Barcode</x-label>
+        
+          
+          </x-input>
+          <x-button id="search-button">
+                <x-icon name="search"></x-icon>
+              </x-button>
+            </div></x-box>      
+    <div id="search-results"></div>
+</body>
+<script type="text/javascript" src="./barcode-search.js"></script>
+<script type="text/javascript" src="./load-barcodes.js"></script>
+<!---
+    <script>var search_barcode = require('./index.js');
+        !-->
+</script>
+
+</html>
