+++ barcode-search.js
@@ -2,8 +2,8 @@
     var dialog = app.dialog;
     var session = app.session;
 
-
-    var defaultBarcode = {"StockNum":"Not Found","CloneName":"Not Found","Box":"Not Found","Position":"Not Found"};
+    var defaultBarcode = {"stockNum":"Not Found","cloneName":"Not Found","box":"Not Found","position":"Not Found"};
+    var results = 'Hello Vue!';
 
     var appResults = new Vue({
         el: '#results-box',
@@ -16,36 +16,32 @@
         }
       });
 
-
-    // search box element
-    let searchBox = document.getElementById("search-box");
-
     function test_func() {
-        searchBox.value = "0055122673"
+        document.querySelector("#search-box").value = document.getElementById("sample-barcode").textContent; //"0055122673" 
+        
         search_func();
     }
 
-    function searchByBarcode(searchCode) { 
-        // open popup warning if no code was submitted
-        if (searchCode === "") {
+    function search_func() {
+        var searchCode = document.querySelector("#search-box").value;
+        console.log(searchCode);
+        if (searchCode == "") {
             let dialog2 = document.querySelector("x-dialog");
             dialog2.opened = true;
         } else {
-        //get results
+        console.log("Searching")
         results = store.get('barcode_data.' + searchCode);
-        // set default "Not Found" values if barcode is found in data
         if (!results) {
             results = defaultBarcode;
         }
-
-        for (var key in results) {
-            if (results.hasOwnProperty(key)) {
-                console.log(key, results[key]);
-                appResults[key] = jsonObj[key];
+            for (var key in results) {
+                if (results.hasOwnProperty(key)) {
+                    console.log(key, results[key]);
+                    appResults[key] = results[key];
+                }
+                
             }
-
-        }
-        // reset focus to search box 
+            
         document.getElementById('search-box').value = "";
         document.getElementById('search-box').focus();
         // save search results to history
@@ -65,8 +61,9 @@
                     'number of barcodes': num_records}
     }
 
-    searchBox.addEventListener('keypress', function (e) {
+    document.getElementById("search-box").addEventListener('click', search_func);
+    document.getElementById("search-box").addEventListener('keypress', function (e) {
         if (e.keyCode === 13) {
-            searchByBarcode(searchBox.value);
+            search_func();
         }
     });
